dt-model-dlt-security-main/requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
base58==2.1.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
bitarray==2.8.1
bleach==6.1.0
blinker==1.6.3
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
component==0.0.1
cryptography==3.4.8
cupshelpers==1.0
cytoolz==0.12.2
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.1.3
executing==2.0.0
fasteners==0.14.1
fastjsonschema==2.18.1
Flask==3.0.0
Flask-Classy==0.6.10
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
future==0.18.2
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gyp==0.1
h5py==3.9.0
hexbytes==0.3.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipfshttpclient==0.8.0a2
ipykernel==6.25.2
ipython==8.16.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras==2.14.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
lru-dict==1.2.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.7
multiaddr==0.0.9
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
netaddr==0.9.0
netifaces==0.11.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.0
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
paramiko==2.9.3
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-solc-x==1.1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.19.0
pycups==2.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
pytz==2022.1
pyunormalize==15.0.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==2.0.1
rpds-py==0.10.3
rsa==4.9
SecretStorage==3.3.1
semantic-version==2.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.11.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
varint==1.0.2
wadllib==1.3.6
wcwidth==0.2.8
web3==5.31.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
